Move to increase cohesion
